refactor: move hooks to another directory
